requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
langchain==0.1.0
langchain-openai==0.0.2
python-dotenv==1.0.0
httpx==0.25.2
sqlalchemy==2.0.23
asyncpg==0.29.0
pgvector==0.2.4
alembic==1.13.1